refactor(aria2): split RPC client responsibilities
