+++ docs/maintenance.md
@@ -74,6 +74,8 @@ Gate 8 result: `DownKyi` is a one-file bootstrap, `DownKyi.Desktop` owns App/UI/
 
 Gate 9 logging local result: Application owns diagnostic contracts and Infrastructure owns the private NLog 6.1.4 provider, bounded sink, redactor, recent buffer, retention and file-backed exporter; Core contains no logging implementation. A 400-record concurrent flush test first exposed whole-configuration loss, and the complete solution later exposed a final-entry disposal race that focused tests had missed. The final sink keeps its async wrapper alive, defers concurrent producers through a bounded barrier, and resets only `FileTarget` handles. Five focused logging rounds, ten complete Infrastructure rounds and all 616 solution tests passed. Strict `AnalysisMode=All` Release build had zero warnings/errors; format changed 0/800 files; module boundaries and vulnerable/deprecated package audits passed; Gitleaks found zero findings across 935 candidate files. The final same-machine 10,000-event benchmark wrote every event with zero drops; its slower flush, larger allocation and one slower producer run remain explicit non-gating evidence.
 
+Gate 9 aria RPC client local result: source history classifies `AriaClient` as hand-maintained DownKyi protocol code first added by commit `587fcfb`, not generated output or a separately synchronized package. The 1,119-line owner is now six responsibility partials between 65 and 333 lines, and the oversized production-file baseline is empty. A deterministic test invokes all 36 public RPC methods and fixes their JSON-RPC method name, ID and token-placement contracts. It exposed and fixed a pre-existing `ChangeUriAsync` defect that sent `aria2.changePosition` instead of `aria2.changeUri`. Strict `AnalysisMode=All` Release build completed with zero warnings/errors; all 713 tests passed across seven test projects; format changed 0/849 files; module boundaries, vulnerable/deprecated package audits and `git diff --check` passed; Gitleaks found zero findings across 985 candidate files. Remote Windows/Linux/macOS quality and CodeQL gates remain required before integration.
+
 PR 07-15 result: Release build completed with zero warnings, 161 tests passed including real FFmpeg/ffprobe seek validation and Host smoke without Prism global container state, format verification passed, and both vulnerable and deprecated package audits were clean. Cross-RID Release builds passed for Windows x86, Linux x64/arm64, and macOS x64/arm64. An isolated Windows process smoke created the main window, accepted close, and exited with code 0 without reading or writing real user data. Native Linux/macOS execution remains owned by their CI runners.
 
 ## Settings Persistence Policy
